+++ include/camera.h
@@ -1,0 +1,45 @@
+#pragma once
+
+#include <glm/glm.hpp>
+#include <glm/gtc/matrix_transform.hpp>
+
+enum class CameraMovement {
+    FORWARD,
+    BACKWARD,
+    LEFT,
+    RIGHT,
+    UP,
+    DOWN
+};
+
+class Camera {
+public:
+    Camera(glm::vec3 position = glm::vec3(0.0f, 0.0f, 1.0f),
+        glm::vec3 up = glm::vec3(0.0f, 1.0f, 0.0f),
+        float yaw = -90.0f,
+        float pitch = 0.0f);
+
+    void ProcessKeyboard(CameraMovement direction, float deltaTime);
+    void ProcessMouseMovement(float xoffset, float yoffset, bool constrainPitch = true);
+    void ProcessMouseScroll(float yoffset);
+    void PrintValues();
+
+    glm::mat4 GetViewMatrix();
+    glm::mat4 GetProjMatrix();
+    float GetZoom();
+
+    glm::vec3 Position;
+    glm::vec3 Front;
+    glm::vec3 Up;
+    glm::vec3 Right;
+private:
+    void updateCameraVectors();
+
+    glm::vec3 WorldUp;
+
+    float Yaw;
+    float Pitch;
+    float MovementSpeed;
+    float MouseSensitivity;
+    float Zoom;
+};