+++ include/scene_manager.h
@@ -1,0 +1,11 @@
+#include <iostream>
+#include <ecs.h>
+
+namespace SceneManager
+{
+
+inline Scene* currentScene;
+
+void SetCurrentScene(Scene& scene);
+
+}