+++ src/scene_manager.cpp
@@ -1,0 +1,11 @@
+#include <scene_manager.h>
+
+namespace SceneManager
+{
+
+void SetCurrentScene(Scene& scene)
+{
+	currentScene = &scene;
+}
+
+}